requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.0.3
openpyxl==3.1.2
selenium==4.15.0
smtplib-ssl==1.0.4
html2text==2020.1.16
fake-useragent==1.4.0
python-dotenv==1.0.0
plotly==5.17.0